refactor city pages
